+++ login-server/src/com/aionemu/loginserver/network/gameserver/clientpackets/CM_HDDBAN_CONTROL.java
@@ -14,7 +14,7 @@ public class CM_HDDBAN_CONTROL extends GsClientPacket {
 
 	@Override
 	protected void readImpl() {
-		type = (byte) readC();
+		type = readC();
 		address = readS();
 		time = readQ();
 	}
